added ability to toggle between celsius and fahrenheit
removed settings option from toolbar menu (seemed unneccessary to include it)

+++ app/src/main/java/kennywong/me/weatherapp/CurrentWeather.java
@@ -22,8 +22,8 @@ public class CurrentWeather {
         public double temp;
         public int pressure;
         public int humidity;
-        public int temp_min;
-        public int temp_max;
+        public double temp_min;
+        public double temp_max;
     }
 
     public List<Weather> getWeather() {
